'权限更改'

+++ src/main/java/com/classManage/tusdt/controller/ClassroomInfoController.java
@@ -2,6 +2,8 @@ package com.classManage.tusdt.controller;
 
 import com.classManage.tusdt.base.common.ResponseData;
 import com.classManage.tusdt.base.constants.Response;
+import com.classManage.tusdt.constants.CommonConstant;
+import com.classManage.tusdt.model.BO.ClassroomNameBO;
 import com.classManage.tusdt.model.ClassroomInfo;
 import com.classManage.tusdt.service.ClassroomInfoService;
 import io.swagger.annotations.*;
@@ -9,6 +11,8 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Description:
@@ -33,11 +37,19 @@ public class ClassroomInfoController {
     )
     @RequestMapping(value = "/addClassroom", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseData<String> addClassroom(@RequestBody ClassroomInfo classroomInfo) {
+    public ResponseData<String> addClassroom(HttpServletRequest request,@RequestBody ClassroomInfo classroomInfo) {
 
-        ResponseData<String> responseData;
-        responseData = classroomInfoService.addClassroom(classroomInfo);
+        ResponseData<String> responseData = new ResponseData<>();
+        Integer schoolId = (Integer) request.getAttribute("schoolId");
+        classroomInfo.setSchoolId(schoolId);
+        Integer level = (Integer) request.getAttribute("level");
+        if (CommonConstant.USER_LEVEL_UNI_ADMIN.equals(level)) {
+            responseData = classroomInfoService.addClassroom(classroomInfo);
+        } else {
+            responseData.setError("权限不足，仅高校管理员可添加教室");
+        }
         return responseData;
+
     }
 
     @ApiOperation(value = "删除一间教室", notes = "删除教室")
@@ -53,7 +65,12 @@ public class ClassroomInfoController {
                                            @RequestParam(value = "roomId",required = true) Integer roomId) {
 
         ResponseData<String> responseData = new ResponseData<>();
-        responseData = classroomInfoService.removeClassRoom(roomId);
+        Integer level = (Integer) request.getAttribute("level");
+        if (CommonConstant.USER_LEVEL_UNI_ADMIN.equals(level)) {
+            responseData = classroomInfoService.removeClassRoom(roomId);
+        } else {
+            responseData.setError("权限不足，仅高校管理员可删除教室");
+        }
         return responseData;
     }
 
@@ -66,10 +83,32 @@ public class ClassroomInfoController {
     )
     @RequestMapping(value = "/modifyRoomInfo", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseData<String> modifyRoomInfo(@RequestBody ClassroomInfo classroomInfo) {
+    public ResponseData<String> modifyRoomInfo(HttpServletRequest request, @RequestBody ClassroomInfo classroomInfo) {
 
-        ResponseData<String> responseData;
-        responseData = classroomInfoService.modifyClassroom(classroomInfo);
+        ResponseData<String> responseData = new ResponseData<>();
+        Integer level = (Integer) request.getAttribute("level");
+        if (CommonConstant.USER_LEVEL_UNI_ADMIN.equals(level)) {
+            responseData = classroomInfoService.modifyClassroom(classroomInfo);
+        } else {
+            responseData.setError("权限不足，仅高校管理员可编辑教室");
+        }
+        return responseData;
+    }
+    @ApiOperation(value = "获取教室列表", notes = "多时间通用")
+    @ApiResponses({@ApiResponse(code = Response.OK, message = "查询成功"),})
+    @ApiImplicitParams(
+            value = {
+                    @ApiImplicitParam(paramType = "header", name = "token", dataType = "String", required = true, value = "token"),
+            }
+    )
+    @RequestMapping(value = "/classroomList", method = RequestMethod.GET)
+    @ResponseBody
+    public ResponseData<List<ClassroomInfo>> singleTimeRoom(HttpServletRequest request) {
+
+        ResponseData<List<ClassroomInfo>> responseData = new ResponseData<>();
+        Integer schoolId = (Integer) request.getAttribute("schoolId");
+        List<ClassroomInfo> classroomInfoList = classroomInfoService.getAllClassroomName(schoolId);
+        responseData.setOK("获取成功",classroomInfoList);
         return responseData;
     }
